refactor: rename interop folder
