+++ FluentSensors/Core/Taskbar/WinTaskbarInfo.cs
@@ -2,7 +2,7 @@ using System;
 using Windows.Graphics;
 
 
-namespace FluentSensors.Core.Interop
+namespace FluentSensors.Core.Taskbar
 {
     // which side of the screen a taskbar is docked to; mirrors the raw ABE_LEFT/TOP/RIGHT/BOTTOM values
     // SHAppBarMessage reports
@@ -17,7 +17,7 @@ namespace FluentSensors.Core.Interop
     // one taskbar as discovered on a poll tick, the primary one or a secondary one on another monitor
     // a plain snapshot, not a live-updating object; two of these compare equal via normal record equality, thats
     // how WinTaskbarService decides whether anything actually changed since the previous tick
-    public record TaskbarInfo(
+    public record WinTaskbarInfo(
         IntPtr Hwnd,
 
         // raw GetWindowRect bounds, includes the invisible DWM frame margin; see WinTaskbarUiaProbe (next phase)
